LA-585 Refactor Upload Request Group - Merge Add recipient
