+++ lib/presentation/widget/upload_request_group/active_closed/active_closed_upload_request_group_viewmodel.dart
@@ -63,7 +63,7 @@ class ActiveClosedUploadRequestGroupViewModel extends UploadRequestGroupTabViewM
 		this._getSorterInteractor,
     this._saveSorterInteractor, 
 		this._sortInteractor, 
-		this._searchUploadRequestGroupsInteractor) : super(store) {
+		this._searchUploadRequestGroupsInteractor) : super(store, _appNavigation) {
 			_storeStreamSubscription = store.onChange.listen((event) {
 				event.uploadRequestGroupState.viewState.fold((failure) => null, (success) {
 					if (success is SearchUploadRequestGroupsNewQuery && event.uiState.searchState.searchStatus == SearchStatus.ACTIVE) {
