+++ lib/presentation/widget/upload_request_group/upload_request_group_tab_common_viewmodel.dart
@@ -33,22 +33,31 @@ import 'package:dartz/dartz.dart';
 import 'package:domain/domain.dart';
 import 'package:flutter/widgets.dart';
 import 'package:linshare_flutter_app/presentation/localizations/app_localizations.dart';
+import 'package:linshare_flutter_app/presentation/model/file/upload_request_presentation_file.dart';
 import 'package:linshare_flutter_app/presentation/redux/actions/ui_action.dart';
 import 'package:linshare_flutter_app/presentation/redux/actions/upload_request_group_action.dart';
 import 'package:linshare_flutter_app/presentation/redux/online_thunk_action.dart';
 import 'package:linshare_flutter_app/presentation/redux/states/app_state.dart';
 import 'package:linshare_flutter_app/presentation/redux/states/ui_state.dart';
+import 'package:linshare_flutter_app/presentation/util/router/app_navigation.dart';
 import 'package:linshare_flutter_app/presentation/util/router/route_paths.dart';
 import 'package:linshare_flutter_app/presentation/view/context_menu/context_menu_builder.dart';
+import 'package:linshare_flutter_app/presentation/view/header/context_menu_header_builder.dart';
 import 'package:linshare_flutter_app/presentation/view/header/simple_bottom_sheet_header_builder.dart';
 import 'package:linshare_flutter_app/presentation/view/order_by/order_by_dialog_bottom_sheet.dart';
 import 'package:linshare_flutter_app/presentation/widget/base/base_viewmodel.dart';
+import 'package:linshare_flutter_app/presentation/widget/upload_request_group_add_recipient/add_recipients_upload_request_group_arguments.dart';
 import 'package:redux/src/store.dart';
+import 'package:redux_thunk/redux_thunk.dart';
 
 typedef OnSelectedSorter = void Function(Sorter sorter);
 
 abstract class UploadRequestGroupTabViewModel extends BaseViewModel {
-  UploadRequestGroupTabViewModel(Store<AppState> store) : super(store);
+  final AppNavigation _appNavigation;
+
+  UploadRequestGroupTabViewModel(
+    Store<AppState> store,
+    this._appNavigation) : super(store);
 
   void openPopupMenuSorter(BuildContext context, Sorter currentSorter, OnSelectedSorter selectSorterAction) {
     ContextMenuBuilder(context)
@@ -68,6 +77,35 @@ abstract class UploadRequestGroupTabViewModel extends BaseViewModel {
 
   bool isInSearchState() {
     return store.state.uiState.isInSearchState();
+  }
+
+  void openContextMenu(BuildContext context, UploadRequestGroup uploadRequestGroup, List<Widget> actionTiles, {Widget? footerAction}) {
+    store.dispatch(_handleContextMenuAction(context, uploadRequestGroup, actionTiles, footerAction: footerAction));
+  }
+
+  ThunkAction<AppState> _handleContextMenuAction(
+      BuildContext context,
+      UploadRequestGroup uploadRequestGroup,
+      List<Widget> actionTiles,
+      {Widget? footerAction}) {
+    return (Store<AppState> store) async {
+      ContextMenuBuilder(context)
+          .addHeader(ContextMenuHeaderBuilder(
+          Key('context_menu_header'),
+          UploadRequestGroupPresentationFile.fromUploadRequestGroup(uploadRequestGroup)).build())
+          .addTiles(actionTiles)
+          .addFooter(footerAction ?? SizedBox.shrink())
+          .build();
+      store.dispatch(UploadRequestGroupAction(Right(UploadRequestGroupContextMenuItemViewState(uploadRequestGroup))));
+    };
+  }
+
+  void goToAddRecipients(UploadRequestGroup request) {
+    _appNavigation.popBack();
+    _appNavigation.push(
+      RoutePaths.addRecipientsUploadRequestGroup,
+      arguments: AddRecipientsUploadRequestGroupArgument(request),
+    );
   }
 
 }
